fix: print visitor org
